Fix focus being removed from the text input widget
Summary:
Each time a message is received, the focus was stolen from
the text input widget, which made it nearly impossible to write
some command to mycroft. This way we only remove the focus from
the widget when we really change its text.

Test Plan:
Write some text and notice the focus is constantly being removed
from the widget when the animation on the top of the plasmoid is animated.

After applying the patch, it allows to write text as expected.

Reviewers: Aiix

Reviewed By: Aiix

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10752